requirements.txt
pandas
spacy
matplotlib
networkx
bertopic
transformers
tweepy
python-dotenv
yt_dlp
